+++ src/vscodeFunctions.ts
@@ -20,24 +20,25 @@ export function createFiles() {
 
 	const launchJson =
 		`{
-			"version": "0.2.0",
-			"configurations": [
-				{
-					"name": "QB64 Build and Run",
-					"type": "QB64",
-					"request": "launch",					
-					"command": "` + "${config:qb64.installPath}/qb64.exe -c " + String.raw`\"` + "${fileDirname}/${fileBasename}" + String.raw`\"` + " -o " + String.raw`\"` + "${fileDirname}/${fileBasenameNoExtension}.exe" + String.raw`\"` + " -x; start " + String.raw`\"` + "${fileDirname}/${fileBasenameNoExtension}.exe" + String.raw`\"` + "\","
-		+ `			"terminalName": "QB64",
-					"terminalIndex": -1, 
-					"showTerminal": true,
-					"linux": {
-						"command": "` + "${config:qb64.installPath}/qb64 -c '${fileDirname}/${fileBasename}' -o '${fileDirname}/${fileBasenameNoExtension}' -x; ${fileDirname}/${fileBasenameNoExtension}\""
+		"version": "0.2.0",
+		"configurations": [
+			{
+				"name": "QB64 Build and Run",
+				"type": "QB64",
+				"request": "launch",					
+				"command": "` + "${config:qb64.installPath}/qb64.exe -c " + String.raw`\"` + "${fileDirname}/${fileBasename}" + String.raw`\"` + " -o " + String.raw`\"` + "${fileDirname}/${fileBasenameNoExtension}.exe" + String.raw`\"` + " -x; start " + String.raw`\"` + "${fileDirname}/${fileBasenameNoExtension}.exe" + String.raw`\"` + "\"," + "\n\t"
+		+ `		"terminalName": "QB64",
+				"terminalIndex": -1, 
+				"showTerminal": true,
+				"linux": {
+					"command": "` + "${config:qb64.installPath}/qb64 '${fileDirname}/${fileBasename}' -c -x -o '${fileDirname}/${fileBasenameNoExtension}'; mv '${fileDirname}/${fileBasenameNoExtension}' '${fileDirname}/${fileBasenameNoExtension}.run'; '${fileDirname}/${fileBasenameNoExtension}.run'\"" + ",\n"
 		+ `		},
-					"osx": {
-						"command": "` + "${config:qb64.installPath}/qb64 -c '${fileDirname}/${fileBasename}' -o '${fileDirname}/${fileBasenameNoExtension}' -x; ${fileDirname}/${fileBasenameNoExtension}\""
+				"osx": {
+					"command": "` + "${config:qb64.installPath}/qb64 '${fileDirname}/${fileBasename}' -c -x -o '${fileDirname}/${fileBasenameNoExtension}'; mv '${fileDirname}/${fileBasenameNoExtension}' '${fileDirname}/${fileBasenameNoExtension}.run'; '${fileDirname}/${fileBasenameNoExtension}.run'\"" + ",\n"
 		+ `		}
-				} 
-			]\n}`;
+			} 
+		]
+	}`;
 
 	let outputChannnel: any = logFunctions.getChannel(logFunctions.channelType.vscode);
 	try {